Add NumerologyChart class and generate LifePath number.
Created the skeleton for the NumerologyChart which will
be the template from creating, storing, retreiving, and
printing a numerology chart based on a date(birthdate)
and person's name. Added methods to generate the lifepath
number by reducing the date using standard methods
which take into account master numbers and karmic
numbers if they're requested.